chore: BWT structure is divided into module
